% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtdata}
\alias{rtdata}
\title{Emotional something reaction time data}
\format{
A data frame with 238 rows and 7 variables:
\describe{
  \item{gender}{(Binary) gender of participant}
  \item{item_1}{Logarithmic reaction time for item 1}
  \item{item_2}{Logarithmic reaction time for item 2}
  \item{item_3}{Logarithmic reaction time for item 3}
  \item{item_4}{Logarithmic reaction time for item 4}
  \item{item_5}{Logarithmic reaction time for item 5}
  \item{item_6}{Logarithmic reaction time for item 6}
}
}
\source{
\url{http://www3.hogrefe.de/buecher/lehrbuecher/psychlehrbuchplus/lehrbuecher/testtheorie-und-testkonstruktion/zusatzmaterial/}
}
\usage{
rtdata
}
\description{
Data from a test to evaluate emotional something something.
}
\keyword{datasets}
