% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCTT-package.R
\docType{package}
\name{shinyCTT-package}
\alias{shinyCTT}
\alias{shinyCTT-package}
\title{shinyCTT: Easy to use analysis tools for psychometric tests in the CTT framework}
\description{
The package provides a Shiny App to analyze datasets in the framework of classical test theory based on the lavaan package. Most importantly, the model assumptions of the 5 CTT models are checked in decreasing order of complexity (tau-congeneric, essentially tau-equivalent, tau-equivalent and essentially tau-parallel, and tau-parallel) and the reliability estimators are adjusted accordingly. Moreover, it provides standard errors and confidence intervals for all parameters, including item reliabilities and error variances.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abeeisnotabug/shinyCTT}
  \item Report bugs at \url{https://github.com/abeeisnotabug/shinyCTT/issues}
}

}
\author{
\strong{Maintainer}: R. Maximilian Bee \email{richard.maximilian.bee@uni-jena.de}

Authors:
\itemize{
  \item Michael Eid \email{michael.eid@fu-berlin.de}
}

}
\keyword{internal}
