% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtdata}
\alias{rtdata}
\title{Emotional something reaction time data}
\format{A data frame with 238 rows and 7 variables:
\describe{
  \item{geschl}{Gender of participant}
  \item{item_1:item_6}{Logarithmic reaction times}
  ...
}}
\source{
\url{http://www3.hogrefe.de/buecher/lehrbuecher/psychlehrbuchplus/lehrbuecher/testtheorie-und-testkonstruktion/zusatzmaterial/}
}
\usage{
rtdata
}
\description{
Data from a test to evaluate emotional something something.
}
\keyword{datasets}
