% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCTTApp.R
\name{shinyCTTApp}
\alias{shinyCTTApp}
\title{A wrapper function for the Shiny App. This is the function that should be called to run the app. It has no arguments.}
\usage{
shinyCTTApp()
}
\value{
None. This function is called for its side effect of launching the Shiny app.
}
\description{
A wrapper function for the Shiny App. This is the function that should be called to run the app. It has no arguments.
}
\examples{
if (interactive()) shinyCTTApp()
}
